+++ src/base/color.js
@@ -2,23 +2,24 @@ const { lightness, saturation } = require('./util');
 
 const color = {
     bg: '#2d313f',
-    fg: '#b8bccd', // lightness('#2d313f', 56),
-    fgDimmed: '#616884',
-    fgBright: '#d0d3dd', // lightness('#2d313f', 64),
+    fgBright: lightness('#2d313f', 64),
+    fg: saturation(lightness('#2d313f', 48), 8),
+    fgSubtle: lightness('#2d313f', 40),
+    fgMuted: lightness('#2d313f', 24),
     red: '#e98989',
-    green: '#7aa89f',
+    green: '#76a99f',
     orange: '#de9e77',
     yellow: '#ecbfa1',
     blue: '#91acd1',
     magenta: '#ada0d3',
-    cyan: '#7fb4ca',
+    cyan: '#7ab5cc',
     cyanBright: '#86e1fc',
-    special: '#de9e77'
+    special: '#de9e77',
 }
 
 const syntax = {
     string: color.green,
-    comment: '#535a71', // lightness(color.bg, 18),
+    comment: color.fgMuted,
     number: color.magenta,
     languageConstant: color.yellow,
     constant: color.fgBright,